Assume maintenance file exists during initialisation
If we set cpuData.isActive to true by default, on its first invocation
we might report CPU idle percentages instead of the appropriate server
state.

This would be noticeable without socket activation only when the first
query happens before the first second, but *with* socket activation
(which is the only supported mode for now) this is happening all the
time on its very first activation, as the program gets started on demand
and only then the first second is starting to run.